+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.5.0
 
 * create new session when referrer or UTM parameter changes
+* update session on event and set bounced to false (keep everything else)
 * fixed User-Agent blacklist (must be lowercase)
 * updated dependencies